Add an option for CSV output (#35)
* create an output function

Instead of printing directly the output, create a function that gets the data
and prints it out.  This will allow for other output formats in the future.

* Add option for CSV output

If the -c argument is given, the output will be in CSV format.

Closes: #32

* Replace the use of an $@ array into a $* string

While not problematic at this point, the previous version of this code
was (correctly) triggering SC2145.

More info: https://github.com/koalaman/shellcheck/wiki/SC2145 .

* CSV output: envelope reason in quotes

The reason message might contain a comma, which would invalidate the
output. Enveloping it within double quotes takes care of that
scenario.  For now, we are assuming no double quotes will exist within
the message itself.